Update readme.md - different env in ios (#281)
The use of /tmp/envfile is a bad idea, especially when there is one machine doing multiple builds on the same machine in parallel (eg. build server). 

IMHO it is much better to copy the correct .env.staging file to the default .env file. In this way we are not modifying system-wide files like /tmp/envfile.

Do you guys agree?

Co-authored-by: luancurti <luancurti@gmail.com>