Don't specify buildToolsVersion on Android (#477)
Since we're depending on a version of the Android Gradle plugin higher
than 3, we don't need to specify a buildToolsVersion. Trying to grab
this value from the root project, in a project that also doesn't specify
it, would fail. Since the only reason we were doing the fetch was to
specify our own buildToolsVersion, I think we can safely remove it.